+++ src/config.rs
@@ -26,7 +26,7 @@
 //!
 //! - [ ] (Not tested) Apple (`apple.com`)
 //! - [ ] (Not tested) Apple Game Center (`gc.apple.com`)
-//! - [ ] (Not tested) Facebook (`facebook.com`)
+//! - [x] Facebook (`facebook.com`)
 //! - [x] GitHub (`github.com`)
 //! - [x] Google (`google.com`)
 //! - [ ] (Not tested) Google Play Games (`playgames.google.com`)