+++ src/oauth/idp.rs
@@ -1,2 +1,3 @@
+pub(super) mod facebook;
 pub(super) mod github;
 pub(super) mod google;